pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "lightspeed-agentic"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn>=0.30.0",
    "pyyaml>=6.0",
    "openai-agents>=0.0.3",
    "llama-stack>=0.7.0",
    "chardet>=7.4.0.post2",
    "pypdf>=6.9.2",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[project.scripts]
lightspeed-agentic = "src.lightspeed_agentic:main"
